LayoutTests/fast/images/text-recognition/mac/layout-for-vertical-writing-mode-expected.txt

PASS writingMode is "vertical-rl"
PASS successfullyParsed is true

TEST COMPLETE
